fix(conflicts): resolve merge conflicts 2
