+++ pkg/manifest/flags.go
@@ -1,0 +1,10 @@
+package manifest
+
+// Flag represents all of the manifest parameters that can be set with explicit
+// flags. Consumers should bind their flag values to these fields directly.
+type Flag struct {
+	Name        string
+	Description string
+	Authors     []string
+	ServiceID   string
+}
